+++ src/gui/map_tab_widget.h
@@ -194,7 +194,7 @@ public:
 
   QSize minimumSizeHint() const override;
 
-  MapView *getMapView(size_t index) const;
+  MapView *getMapView(int index) const;
   MapView *currentMapView() const;
 
   int addTabWithButton(QWidget *widget, const QString &text, QVariant data = QVariant());
